# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.AppCode:
    enum:
    - 0
    - 1000
    - 1001
    - 1002
    - 1003
    - 1004
    - 1005
    - 1006
    - 1007
    type: integer
    x-enum-comments:
      CodeAuthRequired: Request requires authentication; no valid token provided.
      CodeBusinessLogic: Valid request but invalid in the current context.
      CodeConflict: Duplicate resource or incompatible state.
      CodeInternalError: Unexpected server-side error.
      CodeNotFound: Requested resource does not exist.
      CodePermissionDenied: Authenticated but lacks permission to perform this action.
      CodeRateLimited: Too many requests; client should back off.
      CodeSuccess: Request processed successfully.
      CodeValidationError: 'Invalid input: missing field, wrong type, or format violation.'
    x-enum-descriptions:
    - Request processed successfully.
    - 'Invalid input: missing field, wrong type, or format violation.'
    - Request requires authentication; no valid token provided.
    - Authenticated but lacks permission to perform this action.
    - Requested resource does not exist.
    - Duplicate resource or incompatible state.
    - Valid request but invalid in the current context.
    - Too many requests; client should back off.
    - Unexpected server-side error.
    x-enum-varnames:
    - CodeSuccess
    - CodeValidationError
    - CodeAuthRequired
    - CodePermissionDenied
    - CodeNotFound
    - CodeConflict
    - CodeBusinessLogic
    - CodeRateLimited
    - CodeInternalError
...
  dto.CreateLevelBundleRequest:
    properties:
      bundle:
        type: object
      level_id:
        type: string
    required:
    - bundle
    - level_id
    type: object
...
  dto.LevelBundleResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      level_id:
        type: string
      snapshot_id:
        type: string
      state:
        type: integer
      updated_at:
        type: string
      venue_id:
        type: string
    type: object
...
  dto.Response:
    properties:
      code:
        $ref: '#/definitions/dto.AppCode'
      data: {}
      messages:
        items:
          type: string
        type: array
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Digimap backend REST API
  title: Digimap Backend API
  version: "1.0"
paths:
...
  /venues/{id}/snapshots/{snapshotID}/bundles:
    get:
      description: List level bundles for a snapshot (system admin only)
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Snapshot ID
        in: path
        name: snapshotID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.LevelBundleResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: List level bundles
      tags:
      - snapshots
    post:
      consumes:
      - application/json
      description: Create a level bundle within a snapshot (system admin only)
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Snapshot ID
        in: path
        name: snapshotID
        required: true
        type: string
      - description: Bundle details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateLevelBundleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.LevelBundleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Create level bundle
      tags:
      - snapshots
  /venues/{id}/snapshots/{snapshotID}/bundles/{bundleID}:
    delete:
      description: Delete a level bundle (system admin only)
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Snapshot ID
        in: path
        name: snapshotID
        required: true
        type: string
      - description: Bundle ID
        in: path
        name: bundleID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Delete level bundle
      tags:
      - snapshots
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"